requirements.txt
ultralytics
opencv-python
pyyaml
numpy
tqdm
torch
torchvision
pillow
scipy
openvino-dev
